change logic to return result as raw position no more calculated width/height
as this is meant to be used with batch/bash scripting on top, such
scripting will most likely try to find the largest rectangular area thus
logic returning code should not compute width/height immediately
per-individual image basis. Just return the raw position as found from
iterating the pixel data, then it's up to the application to further
work on it will make it work.